requirements.txt
requests
pandas==3.0.0
python-dotenv==1.0.1
resend==2.21.0
urllib3